feat: implement Alloc/Free and chunk access checks
This introduced a new `sym_expr`, `SymPtr` with an `int32` to represent
the base of the ptr (the original allocation address) and a `sym_expr`
to represent the offset of the indexation.

This has proven to be enough to be on feature parity with wasp's
concolic engine (It detects the same vulns).

new_simplify was modified to consider `SymPtr`s for Add and Sub, not
sure which (if any) other operations should be considered.