feat: Add Symmem for classic symbolic execution
This gives an API similar to an HashMap while using a TreeMap under
the hood, the rationale is that this allows Symstatic to not have
to update its binding for symmem when it writes to it.
The public API is relatively large, but the key is having few contact
points with the underlying store, swapping the TreeMap for an HashMap
should be easy requiring only the modification of 5 functions.

Many small things to make Symmem work:
- Symvalue.simplify now merges extracted I32s like extracted I64s
- Symvalue.nland now has a 32 and 64 bit variants